XML -> POJO transformations tested separately
they were buried in, and not testable apart from, common#api,
which made the tests cover too much ground at once and not as
useful as possible in the face of bugs.

this commit tests just the admin side of things, and the tests
are a copy of those for admin#api with file-based rpc function
(tests/200-admin-file.t.coffee).  i'll strip *those* tests
down to mocks in followup commits.

tests for the user side will follow shortly.