preparations for app.configure
there's a partial implementation which just checks the arguments
so far, and the tests don't assert anything beyond that.

the implementation currently recognizes only RAM:, disk:, swap:
(the easy case); i'm leaving the rest for later.

also, it is a copy/paste of app.add to keep the diff simple,
i'll DRY it up in the next commit.