first step towards end-to-end tests
mocha only reports globals introduced inside `it` (in the BDD
interface, the others are surely the same in this regard).
so far, the HTTP verbs from lib/lo/common.coffee were not
reported because the file is require()d in the top-level
scope first (tests/lo/100-rpc.t.coffee).

this is the first time the HTTP verb helper functions are
actually introduced inside `it`, hence the change in the mocha
command line options.