lib/lo/http#request handles binary data
does not assume strings, uses and returns Buffers instead.

the collection of the whole response body is convenient
for current simplistic clients but completely out of place
in library code.
if the library implementation uses streams, it makes no sense
to buffer instead of *exposing* streams.

using Buffers for image downloads would be catastrophic
for instance.